% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mixed_subjects_loss}
\alias{mixed_subjects_loss}
\title{Loss function for parameter estimation. Generally not called directly by users.}
\usage{
mixed_subjects_loss(pars, q_observed, q_predicted, q_llm, lambda = 0)
}
\arguments{
\item{pars}{A vector of item parameters in slope-intercept form. Passed with discriminations first, location second, in item order.}

\item{q_observed}{A quadrature object constructed from running mixed_subjects_quadrature on LLM-generated item responses}

\item{q_predicted}{A quadrature object constructed from running mixed_subjects_quadrature on predicted item responses for the human calibration sample}

\item{lambda}{The power tuning parameter. Takes values between 0 and 1. When lambda is zero, disregards predicted data completely}
}
\value{
Loss for the mixed subjects estimator at the values of \code{pars}
}
\description{
Loss function for parameter estimation. Generally not called directly by users.
}
