% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mixed_subjects_quadrature}
\alias{mixed_subjects_quadrature}
\title{Converts an item response matrix into quadrature form and estimates a 2PL IRT model with parameters in slope-intercept form}
\usage{
mixed_subjects_quadrature(
  resp,
  N_quad = 10,
  eps = 1e-15,
  iterlim = 1e+05,
  irt_pars = NULL
)
}
\arguments{
\item{resp}{An item response matrix in wide form}

\item{N_quad}{The number of quadrature points to compute. Higher numbers can induce numerical errors}

\item{eps}{A tolerance for the minimum values}

\item{iterlim}{Maximum number of Newton-Raphson iterations passed to \code{rmutil::gauss.hermite()}}

\item{irt_pars}{IRT parameters from human calibration for rescaling parameters estimated from predicted responses}
}
\value{
A list object with two components. \verb{$quad} contains a dataframe with the expected sample sizes and number of correct responses at each quadrature point. \verb{$irt_pars} contains the parameter estimates from the fitted IRT model that generated the expected counts
}
\description{
Converts an item response matrix into quadrature form and estimates a 2PL IRT model with parameters in slope-intercept form
}
